requirements.txt
streamlit
dashscope
pypdf
requests
beautifulsoup4
trafilatura
python-dotenv
